+++ src/events/startButtonListener.js
@@ -1,0 +1,9 @@
+import startHandler from "../handlers/startHandler.js";
+import { dom } from "../data.js";
+
+const startButtonListener = dom.startButton.addEventListener(
+  "click",
+  startHandler
+);
+
+export default startButtonListener;
