+++ src/events/stopButtonListener.js
@@ -1,0 +1,9 @@
+import stopHandler from "../handlers/stopHandler.js";
+import { dom } from "../data.js";
+
+const stopButtonListener = dom.stopButton.addEventListener(
+  "click",
+  stopHandler
+);
+
+export default stopButtonListener;
