+++ src/handlers/stopHandler.js
@@ -1,0 +1,10 @@
+import { data } from "../data.js";
+
+const stopHandler = () => {
+  if (data.watchInterval) {
+    clearInterval(data.watchInterval);
+    data.watchInterval = null;
+  }
+};
+
+export default stopHandler;
